demonstrate limitation of resources with systemd
The systemd-demo shows how systemd can be employed to limit the
resource consumptions of individual processes.

To demonstrate CPU usage limitation, we use the option CPUQuota in
stress-cpu-starvation.service. Alternatively, CPU usage can be limited
with the option CPUShares. For demonstration purposes, we choose
CPUQuota because it enforces an upper limit even when the overall CPU
load is small. The effect of CPUShares would only be observable
whenever the CPU resources run low.

Change-Id: I056ef7c1cb937850429b64719cce4f609cead2ee